+++ src/idc/api/__init__.py
@@ -3,6 +3,8 @@ from ._fonts import DEFAULT_FONT_FAMILY, load_font, text_size
 from ._data import ImageData, make_list, flatten_list, jpeg_quality, array_to_image, empty_image, save_image
 from ._data import FORMATS, FORMAT_JPEG, FORMAT_PNG, FORMAT_BMP, FORMAT_EXTENSIONS
 from ._data import ensure_grayscale, grayscale_required_info, ensure_binary, binary_required_info
+from ._depth import DepthData, DepthInformation
+from ._device import DEVICES, DEVICE_AUTO, DEVICE_CPU, DEVICE_CUDA
 from ._generator import Generator, SingleVariableGenerator
 from ._imgcls import ImageClassificationData
 from ._imgseg import ImageSegmentationData, ImageSegmentationAnnotations, combine_layers, split_layers
@@ -12,7 +14,7 @@ from ._utils import locate_file, locate_image, load_image_from_bytes, load_image
 from ._utils import load_labels, save_labels, save_labels_csv
 from ._utils import crop_image, pad_image
 from ._utils import safe_deepcopy
-from ._data_types import DATATYPE_IMGCLS, DATATYPE_OBJDET, DATATYPE_IMGSEG, DATATYPES, data_type_to_class
+from ._data_types import DATATYPE_DEPTH, DATATYPE_IMGCLS, DATATYPE_OBJDET, DATATYPE_IMGSEG, DATATYPES, DATATYPES_LONG, data_type_to_class, data_types_help, DataTypeSupporter
 from ._geometry import locatedobjects_to_shapely, shapely_to_locatedobject, locatedobject_polygon_to_shapely, locatedobject_bbox_to_shapely
 from ._geometry import intersect_over_union, COMBINATIONS, INTERSECT, UNION
 from ._geometry import merge_polygons, fit_located_object, fit_layers