+++ src/idc/filter/depth/__init__.py
@@ -1,0 +1,1 @@
+from ._depth_to_grayscale import DepthToGrayscale